fix: AWS secret check fallback + frontend lint errors (closes #2773)
- Workflow: add vars.AWS_ROLE_TO_ASSUME fallback in secret check step
  (change documented in PR body; cannot push .github/workflows/ via MCP)
- Split AuthContext: AuthContext object → contexts/auth.ts
- Split PriceRefreshContext: PriceRefreshContext object → contexts/priceRefresh.ts
- Split RouteContext: useRoute hook → hooks/useRoute.ts
- Fix familyMvpRedirect.test.ts: close missing describe block,
  remove stale 2-argument call-sites that caused parse error
