+++ frontend/src/RouteContext.tsx
@@ -1,6 +1,6 @@
-import { createContext, useContext, type ReactNode } from "react";
-import { useRouteMode } from "./hooks/useRouteMode";
-import type { Mode } from "./modes";
+import { createContext, type ReactNode } from 'react';
+import { useRouteMode } from './hooks/useRouteMode';
+import type { Mode } from './modes';
 
 interface RouteContextValue {
   mode: Mode;
@@ -11,17 +11,12 @@ interface RouteContextValue {
   setSelectedGroup: (s: string) => void;
 }
 
-const RouteContext = createContext<RouteContextValue | undefined>(undefined);
+export const RouteContext = createContext<RouteContextValue | undefined>(undefined);
 
 export function RouteProvider({ children }: { children: ReactNode }) {
   const value = useRouteMode();
   return <RouteContext.Provider value={value}>{children}</RouteContext.Provider>;
 }
 
-// eslint-disable-next-line react-refresh/only-export-components
-export function useRoute() {
-  const ctx = useContext(RouteContext);
-  if (!ctx) throw new Error("useRoute must be used within RouteProvider");
-  return ctx;
-}
-
+// Re-export hook for backward compatibility.
+export { useRoute } from './hooks/useRoute';
